server/requirements.txt
# ---- server/requirements.txt ----
# Flask: Web framework
# Flask-PyMongo: MongoDB integration for Flask
# python-dotenv: For loading environment variables from .env file (good for API keys)
# requests: For making HTTP requests (e.g., to AI APIs)
# Flask-Cors: To handle Cross-Origin Resource Sharing if your frontend is on a different domain/port
# bcrypt: For hashing passwords

Flask
Flask-PyMongo
python-dotenv
requests
Flask-Cors
bcrypt
from flask import Flask, request, jsonify
from flask_cors import CORS
import os
import json
import google.generativeai as genai
from google.generativeai import list_models
